Visitor pattern partial implementation
Attempted to implement visitor pattern but I'm coming up on more
barriers than it solves.  Restoring from DTOs is fine and requires the
pattern to be implemented on the DTOs.  To create from IParameter, these
would also need another implementation of the visitor pattern which
tightly couples the concrete implementation of what criteria exist into
the core library which is unacceptable.  A more suitable pattern is
required.