Added tree selection to output fields
As per the comments in #7, a TreeView is a faster method of adding lots
of fields so this has been added.  Whole subjects can be added by
dragging the subject (or relation field) nodes.  Can't reorder or delete
output fields yet.  Will still need a combo box version as it's needed
in the advanced search.