Auto-use fields on change
To make it faster to use, as soon as you start modifying the output
fields the option for 'Use custom output fields' will auto-select.
Likewise when all output fields have been removed it'll auto uncheck
(not valid to search for no fields).  If the fields aren't being changed
the user is still able to turn the fields on or off.