Refactored parsing
The purpose of the last two commits is to enable we can retrieve raw
values from controls (unaffected by parsing) so we can implement
serialisation and advanced search correctly.

Removed all parsing logic from UIElement and added it to IPartView (of
which PresetPart and StandardPart implement).
Added ListBoxElement to WPF for limited list fields.
Began refactoring ExtendedControlFactory to ParserFactory which will
need to be hooked into a CreatePart override for PresetViewAdapter and
StandardViewAdapter.