Fixed templating issue
The adapter changing an index is what caused the rendering issues in the
templates.  This has changed to a remove then add.
Prefix needs to move to the parent AdvancedPart and should be used in
the Junction template as currently it doesn't show the prefix in front
of a nested junction.