Updated sample with DateParser
Since the SQL ParameterBuilders that deal with dates want DateValues to
process, it's a requirement that we MUST have a parser on the part to
ensure the raw UI value is converted to DateValue for the
ParameterBuilder. So there's now events on the adapters that occur when
a part is created, allowing the consumer to modify the part during
construction.